+++ pkg/annotation/annotation.go
@@ -6,4 +6,5 @@ const (
 	MachineDeploymentSubnet = "machine-deployment.giantswarm.io/subnet"
 	// Annotation used to enable feature to terminate unhealthy nodes on a cluster CR.
 	NodeTerminateUnhealthy = "alpha.node.giantswarm.io/terminate-unhealthy"
+	AWSMetadata             = "alpha.giantswarm.io/aws-metadata-v2"
 )