+++ CHANGELOG.md
@@ -11,6 +11,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add `terminate-unhealthy-node` feature to automaticaly terminate bad and unhealthy nodes in a Cluster.
+- Annotation `alpha.giantswarm.io/aws-metadata-v2` to enable AWS Metadata API v2
+
+### Fixed
+
+- Remove dependency on `var-lib-etcd.automount` to avoid dependency cycle on new systemd.
 
 ## [9.1.3] - 2020-10-21